Refactor ATI params so frontPage and IDX pages use src/app/containers/ATIAnalytics/params/indexPage
