--- src/app/containers/ATIAnalytics/params/frontpage/buildParams.js
@@ -1,50 +1,0 @@
-import { buildATIPageTrackPath } from '../../atiUrl';
-import { getPublishedDatetime, LIBRARY_VERSION } from '#lib/analyticsUtils';
-import {
-  getContentId,
-  getLanguage,
-  getPageIdentifier,
-  getPageTitle,
-} from '#lib/analyticsUtils/frontpage';
-
-export const buildFrontPageATIParams = (
-  frontpageData,
-  requestContext,
-  serviceContext,
-) => {
-  const { platform, statsDestination } = requestContext;
-  const {
-    atiAnalyticsAppName,
-    atiAnalyticsProducerId,
-    brandName,
-    service,
-  } = serviceContext;
-
-  return {
-    appName: atiAnalyticsAppName,
-    contentId: getContentId(frontpageData),
-    contentType: 'index-home',
-    language: getLanguage(frontpageData),
-    pageIdentifier: getPageIdentifier(frontpageData, service),
-    pageTitle: getPageTitle(frontpageData, brandName),
-    producerId: atiAnalyticsProducerId,
-    timePublished: getPublishedDatetime('firstPublished', frontpageData),
-    timeUpdated: getPublishedDatetime('lastPublished', frontpageData),
-    libraryVersion: LIBRARY_VERSION,
-    platform,
-    service,
-    statsDestination,
-  };
-};
-
-export const buildFrontPageATIUrl = (
-  frontpageData,
-  requestContext,
-  serviceContext,
-) => {
-  return buildATIPageTrackPath(
-    buildFrontPageATIParams(frontpageData, requestContext, serviceContext),
-  );
-};
-
-export default buildFrontPageATIParams;
